requirements.txt
Flask
Flask-CORS
Flask-Limiter
SQLAlchemy
psycopg2-binary
redis
python-dotenv
gunicorn
pytest
